Split Approximate
