+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/approximation/Approximation.java
@@ -9,24 +9,16 @@ package org.elasticsearch.xpack.esql.approximation;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.elasticsearch.common.io.stream.StreamOutput;
+import org.apache.lucene.util.SetOnce;
 import org.elasticsearch.common.util.set.Sets;
 import org.elasticsearch.compute.data.LongBlock;
-import org.elasticsearch.core.Tuple;
 import org.elasticsearch.index.IndexMode;
 import org.elasticsearch.xpack.esql.VerificationException;
 import org.elasticsearch.xpack.esql.common.Failure;
 import org.elasticsearch.xpack.esql.core.expression.Alias;
-import org.elasticsearch.xpack.esql.core.expression.Attribute;
-import org.elasticsearch.xpack.esql.core.expression.Expression;
-import org.elasticsearch.xpack.esql.core.expression.LeafExpression;
 import org.elasticsearch.xpack.esql.core.expression.Literal;
-import org.elasticsearch.xpack.esql.core.expression.NameId;
 import org.elasticsearch.xpack.esql.core.expression.NamedExpression;
-import org.elasticsearch.xpack.esql.core.expression.Nullability;
-import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
 import org.elasticsearch.xpack.esql.core.tree.Source;
-import org.elasticsearch.xpack.esql.core.type.DataType;
 import org.elasticsearch.xpack.esql.core.util.Holder;
 import org.elasticsearch.xpack.esql.core.util.StringUtils;
 import org.elasticsearch.xpack.esql.expression.function.aggregate.AggregateFunction;
@@ -39,22 +31,8 @@ import org.elasticsearch.xpack.esql.expression.function.aggregate.StdDev;
 import org.elasticsearch.xpack.esql.expression.function.aggregate.Sum;
 import org.elasticsearch.xpack.esql.expression.function.aggregate.Variance;
 import org.elasticsearch.xpack.esql.expression.function.aggregate.WeightedAvg;
-import org.elasticsearch.xpack.esql.expression.function.scalar.EsqlScalarFunction;
 import org.elasticsearch.xpack.esql.expression.function.scalar.approximate.ConfidenceInterval;
-import org.elasticsearch.xpack.esql.expression.function.scalar.approximate.Random;
-import org.elasticsearch.xpack.esql.expression.function.scalar.conditional.Case;
-import org.elasticsearch.xpack.esql.expression.function.scalar.convert.ToDouble;
-import org.elasticsearch.xpack.esql.expression.function.scalar.convert.ToInteger;
-import org.elasticsearch.xpack.esql.expression.function.scalar.convert.ToLong;
-import org.elasticsearch.xpack.esql.expression.function.scalar.multivalue.MvAppend;
-import org.elasticsearch.xpack.esql.expression.function.scalar.multivalue.MvSlice;
-import org.elasticsearch.xpack.esql.expression.function.scalar.nulls.Coalesce;
-import org.elasticsearch.xpack.esql.expression.predicate.logical.And;
-import org.elasticsearch.xpack.esql.expression.predicate.nulls.IsNotNull;
-import org.elasticsearch.xpack.esql.expression.predicate.nulls.IsNull;
-import org.elasticsearch.xpack.esql.expression.predicate.operator.arithmetic.Div;
-import org.elasticsearch.xpack.esql.expression.predicate.operator.comparison.Equals;
-import org.elasticsearch.xpack.esql.expression.predicate.operator.comparison.GreaterThanOrEqual;
+import org.elasticsearch.xpack.esql.optimizer.rules.logical.SubstituteApproximationPlan;
 import org.elasticsearch.xpack.esql.plan.logical.Aggregate;
 import org.elasticsearch.xpack.esql.plan.logical.ChangePoint;
 import org.elasticsearch.xpack.esql.plan.logical.Dissect;
@@ -79,19 +57,10 @@ import org.elasticsearch.xpack.esql.plan.logical.TopN;
 import org.elasticsearch.xpack.esql.plan.logical.UriParts;
 import org.elasticsearch.xpack.esql.plan.logical.inference.Completion;
 import org.elasticsearch.xpack.esql.plan.logical.inference.Rerank;
-import org.elasticsearch.xpack.esql.plan.logical.local.LocalRelation;
-import org.elasticsearch.xpack.esql.session.Configuration;
 import org.elasticsearch.xpack.esql.session.Result;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /**
  * This class computes approximate results for certain classes of ES|QL queries.
@@ -109,36 +78,40 @@ import java.util.stream.Stream;
  * </ul>
  * Some of these restrictions may be lifted in the future.
  * <p>
- *
  * When these conditions are met, the query is replaced by an approximation query
  * that samples documents before the STATS, and extrapolates the aggregate
  * functions if needed. This new logical plan is generated by
- * {@link Approximation#approximationPlan}.
+ * {@link ApproximationPlan#get}. The substitution of the original query by the
+ * approximation query happens during logical optimization, in the rule
+ * {@link SubstituteApproximationPlan}.
  * <p>
  * In addition to approximate results, confidence intervals are also computed.
- * This is done by dividing the sampled rows {@link Approximation#TRIAL_COUNT}
- * times into {@link Approximation#BUCKET_COUNT} buckets, computing the aggregate
+ * This is done by dividing the sampled rows {@link ApproximationPlan#TRIAL_COUNT}
+ * times into {@link ApproximationPlan#BUCKET_COUNT} buckets, computing the aggregate
  * functions for each bucket, and using these sampled values to compute
- * confidence intervals.
+ * confidence intervals with the bias-corrected and accelerated (BCa) bootstrap
+ * method, see also {@link ConfidenceInterval}.
  * <p>
+ * The initial approximation plan contains a placeholder for the sample probability,
+ * which is determined during subplan execution, and is based on results set size.
  * To obtain an appropriate sample probability, first a target number of rows
- * is set. For now this is determined via the ({@link ApproximationSettings}).
- * Next, the total number of rows in the source index is counted via the plan
- * {@link Approximation#sourceCountPlan}. This plan always executes fast. When
+ * is set. This is determined via the {@link ApproximationSettings}.
+ * Next, the total number of rows in the source index is counted via the subplan
+ * {@link Approximation#sourceCountSubPlan}. This plan always executes fast. When
  * there are no commands that can change the number of rows, the sample
  * probability can be directly computed as a ratio of the target number of rows
  * and this total number.
  * <p>
  * In the presence of commands that can change the number of rows, another step
- * is needed. The first goal is to find a sample probability that leads to
+ * is needed. The first goal is to find a sample probability that leads to approximately
  * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} rows, and when is
  * probability is found, a sample probability leading to the target number of
- * row is computed.
+ * rows is computed.
  * <p>
  * This is done by setting the initial sample probability to the ratio of
  * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} and the total number
- * of rows in the source index, and a number of rows is sampled with the plan
- * {@link Approximation#countPlan}. As long as the sampled number of rows is too
+ * of rows in the source index, and a number of rows is sampled with the subplan
+ * {@link Approximation#countSubPlan}. As long as the sampled number of rows is too
  * small, the probability is scaled up until a good probability is reached. This
  * final probability is used to compute the probability using for approximating
  * the original query.
@@ -250,29 +223,12 @@ public class Approximation {
     );
 
     /**
-     * These aggregate functions need to be corrected for random sampling, by
-     * scaling up the sampled value by the inverse of the sampling probability.
-     * Other aggregate functions do not need any correction.
-     */
-    private static final Set<Class<? extends AggregateFunction>> SAMPLE_CORRECTED_AGGS = Set.of(Count.class, Sum.class);
-
-    /**
      * These multivalued aggregate functions work well with random sampling.
      * However, confidence intervals make no sense anymore and are dropped.
      */
     static final Set<Class<? extends AggregateFunction>> SUPPORTED_MULTIVALUED_AGGS = Set.of(
         org.elasticsearch.xpack.esql.expression.function.aggregate.Sample.class
     );
-
-    /**
-     * These numerical scalar functions produce multivalued output. This means that
-     * confidence intervals cannot be computed anymore and are dropped.
-     * <p>
-     * Numerical scalar functions that produce multivalued output should be added
-     * here. Forgetting to do so leads to confidence intervals columns for the
-     * multivalued fields, that are filled with nulls.
-     */
-    private static final Set<Class<? extends EsqlScalarFunction>> MULTIVALUED_OUTPUT_FUNCTIONS = Set.of(MvAppend.class);
 
     /**
      * The target number of sampled rows to estimate the total number of rows
@@ -295,114 +251,37 @@ public class Approximation {
     private static final int DEFAULT_ROW_COUNT_WITH_GROUPING = 1_000_000;
 
     /**
-     * Default confidence level for confidence intervals.
-     */
-    private static final double DEFAULT_CONFIDENCE_LEVEL = 0.90;
-
-    /**
      * Don't sample with a probability higher than this threshold. The cost of
      * tracking confidence intervals doesn't outweigh the benefits of sampling.
      */
     private static final double SAMPLE_PROBABILITY_THRESHOLD = 0.1;
 
-    /**
-     * The number of times (trials) the sampled rows are divided into buckets.
-     */
-    private static final int TRIAL_COUNT = 2;
-
-    /**
-     * The number of buckets to use for computing confidence intervals.
-     */
-    private static final int BUCKET_COUNT = 15;
-
-    /**
-     * For grouped statistics (STATS ... BY), a grouping needs at least this
-     * number of sampled rows to be included in the results. For a simple
-     * aggregation as count, this still leads to acceptable confidence
-     * intervals. This will never be marked as "certified reliable".
-     */
-    private static final int MIN_ROW_COUNT_FOR_RESULT_INCLUSION = 10;
-
     private static final Logger logger = LogManager.getLogger(Approximation.class);
 
     private static final AggregateFunction COUNT_ALL_ROWS = new Count(Source.EMPTY, Literal.keyword(Source.EMPTY, StringUtils.WILDCARD));
-
-    public static class PlaceHolder extends LeafExpression {
-
-        private final String name;
-
-        PlaceHolder(String name) {
-            super(Source.EMPTY);
-            this.name = name;
-        }
-
-        @Override
-        public Nullability nullable() {
-            return Nullability.TRUE;
-        }
-
-        @Override
-        public DataType dataType() {
-            return DataType.DOUBLE;
-        }
-
-        @Override
-        public String getWriteableName() {
-            throw new UnsupportedOperationException("not serialized");
-        }
-
-        @Override
-        public void writeTo(StreamOutput out) throws IOException {
-            throw new UnsupportedOperationException("not serialized");
-        }
-
-        @Override
-        protected NodeInfo<? extends Expression> info() {
-            return NodeInfo.create(this);
-        }
-
-        @Override
-        public String toString() {
-            return "PlaceHolder[" + name + "]";
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            return obj != null && getClass() == obj.getClass() && Objects.equals(name, ((PlaceHolder) obj).name);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(name);
-        }
-    }
-
-    private static class SampleProbability extends PlaceHolder {
-        SampleProbability() {
-            super("SampleProbability");
-        }
-    }
-
-    private static class SampleSizeThreshold extends PlaceHolder {
-        SampleSizeThreshold() {
-            super("SampleSizeThreshold");
-        }
-    }
 
     private final LogicalPlan logicalPlan;
     private final ApproximationSettings settings;
     private final QueryProperties queryProperties;
 
+    private Double nextSubPlanSampleProbability;
+    private int subPlanIterationCount;
+    private final SetOnce<Long> sourceRowCount;
+
     public Approximation(LogicalPlan logicalPlan, ApproximationSettings settings) {
         this.logicalPlan = logicalPlan;
         this.settings = settings;
         this.queryProperties = verifyPlan(logicalPlan);
+
+        nextSubPlanSampleProbability = null;
+        subPlanIterationCount = 0;
+        sourceRowCount = new SetOnce<>();
     }
 
     /**
      * Verifies that a plan is suitable for approximation.
      *
-     * @return whether the part of the query until the STATS command preserves all rows
+     * @return the query properties relevant for approximation
      * @throws VerificationException if the plan is not suitable for approximation
      */
     public static QueryProperties verifyPlan(LogicalPlan logicalPlan) throws VerificationException {
@@ -462,106 +341,28 @@ public class Approximation {
         return new QueryProperties(hasGrouping.get(), canDecreaseRowCount.get(), canIncreaseRowCount.get());
     }
 
-    public static LogicalPlan firstSubPlan(LogicalPlan logicalPlan, Configuration configuration, Set<LocalRelation> subPlansResults) {
-        Holder<Boolean> hasSampleProbability = new Holder<>(false);
-        logicalPlan.forEachExpressionDown(SampleProbability.class, prob -> hasSampleProbability.set(true));
-        if (hasSampleProbability.get() == false) {
+    /**
+     * Returns the first subplan to execute for approximation, or null if the main plan can be executed directly.
+     */
+    public LogicalPlan firstSubPlan() {
+        if (sourceRowCount.get() == null) {
+            return sourceCountSubPlan();
+        } else if (nextSubPlanSampleProbability != null) {
+            return countSubPlan(nextSubPlanSampleProbability);
+        } else {
             return null;
         }
+    }
 
-        Map<String, Long> rowCounts = subPlansResults.stream()
-            .collect(
-                Collectors.toMap(
-                    localRelation -> localRelation.output().getFirst().name(),
-                    localRelation -> ((LongBlock) localRelation.supplier().get().getBlock(0)).getLong(0)
-                )
-            );
-
-        if (rowCounts.size() == 0) {
-            return sourceCountPlan(logicalPlan);
-        } else if (rowCounts.size() == 1) {
-            return countPlan(logicalPlan, Math.min(1.0, (double) ROW_COUNT_FOR_COUNT_ESTIMATION / rowCounts.get("$source_count")));
+    /**
+     * Returns the new main plan to execute for approximation after executing a subplan, based on the result of the subplan.
+     */
+    public LogicalPlan newMainPlan(Result result) {
+        if (sourceRowCount.get() == null) {
+            return processSourceCount(rowCount(result));
         } else {
-            Tuple<Double, Long> result = getResult(rowCounts);
-            return countPlan(logicalPlan, Math.min(1.0, result.v1() * ROW_COUNT_FOR_COUNT_ESTIMATION / Math.max(1, result.v2())));
+            return processCount(rowCount(result));
         }
-    }
-
-    public static LogicalPlan newMainPlan(LogicalPlan logicalPlan, Configuration configuration, Set<LocalRelation> subPlansResults) {
-        System.out.println("@@newMainPlan: " + subPlansResults);
-
-        Map<String, Long> rowCounts = subPlansResults.stream()
-            .collect(
-                Collectors.toMap(
-                    localRelation -> localRelation.output().getFirst().name(),
-                    localRelation -> ((LongBlock) localRelation.supplier().get().getBlock(0)).getLong(0)
-                )
-            );
-        System.out.println("@@rowCounts: " + rowCounts);
-
-        if (rowCounts.size() == 1) {
-            return processSourceCount(logicalPlan, configuration, rowCounts.get("$source_count"));
-        } else {
-            Tuple<Double, Long> result = getResult(rowCounts);
-            return processCount(logicalPlan, configuration, result.v1(), result.v2());
-        }
-    }
-
-    private static Tuple<Double, Long> getResult(Map<String, Long> rowCounts) {
-        double sampleProbability = rowCounts.keySet()
-            .stream()
-            .filter(k -> k.startsWith("$count_p="))
-            .map(k -> Double.parseDouble(k.substring("$count_p=".length())))
-            .max(Double::compareTo)
-            .orElseThrow(() -> new IllegalStateException("expected a key starting with $count_p=, but got " + rowCounts.keySet()));
-        return Tuple.tuple(sampleProbability, rowCounts.get("$count_p="+sampleProbability));
-
-    }
-
-    private static LogicalPlan substituteSampleProbability(LogicalPlan logicalPlan, double sampleProbability) {
-        logicalPlan = logicalPlan.transformExpressionsDown(
-            SampleProbability.class,
-            prob -> Literal.fromDouble(Source.EMPTY, sampleProbability)
-        );
-        if (sampleProbability == 1.0) {
-            logicalPlan = logicalPlan.transformDown(SampledAggregate.class, agg -> {
-                List<Alias> nullBuckets = new ArrayList<>();
-                Set<String> originalAggs = agg.originalAggregates().stream().map(NamedExpression::name).collect(Collectors.toSet());
-                for (Attribute attr : agg.outputSet()) {
-                    if (originalAggs.contains(attr.name()) == false) {
-                        System.out.println(" *** NULLIFY: " + attr);
-                        nullBuckets.add(new Alias(Source.EMPTY, attr.name(), Literal.NULL, attr.id()));
-                    } else {
-                        System.out.println(" *** KEEP   : " + attr);
-                    }
-                }
-
-                LogicalPlan plan = new Aggregate(agg.source(), agg.child(), agg.groupings(), agg.originalAggregates());
-                plan = new Eval(Source.EMPTY, plan, nullBuckets);
-                return plan;
-            });
-        }
-        logicalPlan = logicalPlan.transformExpressionsDown(
-            SampleSizeThreshold.class,
-            prob -> Literal.integer(Source.EMPTY, sampleProbability == 1.0 ? 1 : MIN_ROW_COUNT_FOR_RESULT_INCLUSION)
-        );
-        logicalPlan.setOptimized();
-        return logicalPlan;
-    }
-
-    private static int sampleRowCount(ApproximationSettings settings) {
-        if (settings.rows() != null) {
-            return settings.rows();
-            // TODO!
-            // } else if (queryProperties.hasGrouping) {
-            // return DEFAULT_ROW_COUNT_WITH_GROUPING;
-        } else {
-            return DEFAULT_ROW_COUNT_WITHOUT_GROUPING;
-        }
-    }
-
-    private double confidenceLevel() {
-        return settings.confidenceLevel() != null ? settings.confidenceLevel() : DEFAULT_CONFIDENCE_LEVEL;
     }
 
     /**
@@ -571,7 +372,7 @@ public class Approximation {
      * {@code FROM index | STATS COUNT(*)}
      * </pre>
      */
-    private static LogicalPlan sourceCountPlan(LogicalPlan logicalPlan) {
+    private LogicalPlan sourceCountSubPlan() {
         LogicalPlan leaf = logicalPlan.collectLeaves().getFirst();
         LogicalPlan sourceCountPlan = new Aggregate(
             Source.EMPTY,
@@ -583,34 +384,34 @@ public class Approximation {
         return sourceCountPlan;
     }
 
-    private static LogicalPlan processSourceCount(LogicalPlan logicalPlan, Configuration configuration, long sourceRowCount) {
+    /**
+     * Determines either the final sample probability or whether more subplans
+     * need to the executed, based on the total number of rows in the source
+     * index and the query properties.
+     */
+    private LogicalPlan processSourceCount(long sourceRowCount) {
         logger.debug("total number of source rows: [{}] rows", sourceRowCount);
+        this.sourceRowCount.set(sourceRowCount);
         if (sourceRowCount == 0) {
             // If there are no rows, run the original query.
-            return substituteSampleProbability(logicalPlan, 1.0);
+            nextSubPlanSampleProbability = null;
+            return ApproximationPlan.substituteSampleProbability(logicalPlan, 1.0);
         }
-        QueryProperties queryProperties = verifyPlan(logicalPlan);
-        double sampleProbability = Math.min(1.0, (double) sampleRowCount(configuration.approximationSettings()) / sourceRowCount);
-        if (queryProperties.canIncreaseRowCount == false && queryProperties.canDecreaseRowCount == false) {
-            // If the query preserves all rows, we can directly approximate with the sample probability.
-            return substituteSampleProbability(logicalPlan, sampleProbability);
-        } else if (queryProperties.canIncreaseRowCount == false && sampleProbability > SAMPLE_PROBABILITY_THRESHOLD) {
+        double sampleProbability = Math.min(1.0, (double) sampleRowCount() / sourceRowCount);
+        if (queryProperties.canIncreaseRowCount == false && sampleProbability > SAMPLE_PROBABILITY_THRESHOLD) {
             // If the query cannot increase the number of rows, and the sample probability is large,
             // we can directly run the original query without sampling.
             logger.debug("using original plan (too few rows)");
-            return substituteSampleProbability(logicalPlan, 1.0);
+            nextSubPlanSampleProbability = null;
+            return ApproximationPlan.substituteSampleProbability(logicalPlan, 1.0);
+        } else if (queryProperties.canIncreaseRowCount == false && queryProperties.canDecreaseRowCount == false) {
+            // If the query preserves all rows, we can directly approximate with the sample probability.
+            nextSubPlanSampleProbability = null;
+            return ApproximationPlan.substituteSampleProbability(logicalPlan, sampleProbability);
         } else {
             // Otherwise, we need to sample the number of rows first to obtain a good sample probability.
+            nextSubPlanSampleProbability = Math.min(1.0, (double) ROW_COUNT_FOR_COUNT_ESTIMATION / sourceRowCount);
             return logicalPlan;
-            // sampleProbability = Math.min(1.0, (double) ROW_COUNT_FOR_COUNT_ESTIMATION / sourceRowCount.get());
-            //
-            // runner.run(
-            // toPhysicalPlan.apply(countPlan(sampleProbability)),
-            // configuration,
-            // foldContext,
-            // planTimeProfile,
-            // countListener(sampleProbability, listener)
-            // );
         }
     }
 
@@ -622,19 +423,20 @@ public class Approximation {
      * </pre>
      * where {@code prob} is the given sample probability.
      */
-    private static LogicalPlan countPlan(LogicalPlan logicalPlan, double sampleProbability) {
+    private LogicalPlan countSubPlan(double sampleProbability) {
         Holder<Boolean> encounteredStats = new Holder<>(false);
         LogicalPlan countPlan = logicalPlan.transformUp(plan -> {
             if (encounteredStats.get() == false) {
                 if (plan instanceof Aggregate aggregate) {
                     // The STATS function should be replaced by a STATS COUNT(*).
                     encounteredStats.set(true);
+                    List<NamedExpression> aggregations = List.of(new Alias(Source.EMPTY, "$count_p=" + sampleProbability, COUNT_ALL_ROWS));
                     plan = new SampledAggregate(
                         Source.EMPTY,
                         aggregate.child(),
                         List.of(),
-                        List.of(new Alias(Source.EMPTY, "$count_p=" + sampleProbability, COUNT_ALL_ROWS)),
-                        List.of(new Alias(Source.EMPTY, "$count_p=" + sampleProbability, COUNT_ALL_ROWS)),
+                        aggregations,
+                        aggregations,
                         Literal.fromDouble(Source.EMPTY, sampleProbability)
                     );
                 }
@@ -649,81 +451,97 @@ public class Approximation {
         return countPlan;
     }
 
-    // /**
-    // * Receives the sampled number of rows reaching the STATS function.
-    // * Runs either the {@link Approximation#approximationPlan} or a next iteration
-    // * {@link Approximation#countPlan} depending on whether the current count is
-    // * sufficient.
-    // * <p>
-    // * This count listener works recursively, increasing the sample probability at
-    // * each step. The first iteration uses a probability of
-    // * <pre>
-    // * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} / {@link Approximation#sourceRowCount}
-    // * </pre>
-    // * which is set by the {@link Approximation#sourceCountListener}.
-    // * <p>
-    // * If the sampled row count is high enough, the approximation plan is run next.
-    // * <p>
-    // * If the sampled row count is low the probability is multiplied by
-    // * <pre>
-    // * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} / Math.max(1, rowCount)
-    // * </pre>
-    // * and the count plan is run again.
-    // * <p>
-    // * For positive rowCount this means the next iteration aims for
-    // * approximately {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} rows.
-    // * The recursion stops when a rowCount larger than {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} / 2
-    // * is encountered. The division by 2 is to stop when the random sampling
-    // * returns a few less rows than aimed for.
-    // * <p>
-    // * For zero rowCount the sample probability is multiplied by
-    // * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION}. That means the next
-    // * recursion step can sample a number of rows between 0 and the target count,
-    // * but can never overshoot by much.
-    // * <p>
-    // * Regarding the recursion depth: in the worst case (very large index, very few
-    // * matching docs), the first few iterations return zero rows, and the sample probability
-    // * is multiplied by {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} each time.
-    // * Next, there's one iteration that returns a positive small row count, and the
-    // * probability is scaled up the correct value. Next, there's one final iteration
-    // * with approximately {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} rows.
-    // * <p>
-    // * This result in a maximum recursion depth of about:
-    // * <pre>
-    // * log(sourceRowCount, {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION}) + 2
-    // * </pre>
-    // * which is at most 7 when sourceRowCount=MAX_LONG (huge!) and
-    // * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} = 10000.
-    // * To be safe, the maximum recursion depth is capped at 10, and an exception is thrown
-    // * when this depth is exceeded.
-    // */
-    private static LogicalPlan processCount(LogicalPlan logicalPlan, Configuration configuration, double sampleProbability, long rowCount) {
-        // countRecursionDepth += 1;
-        // if (countRecursionDepth > 10) {
-        // listener.onFailure(new IllegalStateException("Approximation count recursion limit exceeded"));
-        // return;
-        // }
+     /**
+      * Receives the sampled number of rows reaching the STATS function and determines
+      * whether the next iteration should run the main approximation plan (if a good
+      * sample probability is found), or another {@link Approximation#countSubPlan}
+      * with higher sample probability,
+      * <p>
+      * The first iteration uses a probability of
+      * <pre>
+      * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} / {@link Approximation#sourceRowCount}
+      * </pre>
+      * which is set by the {@link Approximation#processSourceCount}.
+      * <p>
+      * If the sampled row count is high enough, the approximation plan is run next.
+      * <p>
+      * If the sampled row count is low the probability is multiplied by
+      * <pre>
+      * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} / Math.max(1, rowCount)
+      * </pre>
+      * and the count plan is run again.
+      * <p>
+      * For positive rowCount this means the next iteration aims for
+      * approximately {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} rows.
+      * The iteration stops when a rowCount larger than {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} / 2
+      * is encountered. The division by 2 is to stop when the random sampling
+      * returns a few less rows than aimed for.
+      * <p>
+      * For zero rowCount the sample probability is multiplied by
+      * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION}. That means the next
+      * iteration can sample a number of rows between 0 and the target count,
+      * but can never overshoot by much.
+      * <p>
+      * Regarding the iteration count: in the worst case (very large index, very few
+      * matching docs), the first few iterations return zero rows, and the sample probability
+      * is multiplied by {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} each time.
+      * Next, there's one iteration that returns a positive small row count, and the
+      * probability is scaled up the correct value. Next, there's one final iteration
+      * with approximately {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} rows.
+      * <p>
+      * This result in a maximum iteration count of about:
+      * <pre>
+      * log(sourceRowCount, {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION}) + 2
+      * </pre>
+      * which is at most 7 when sourceRowCount=MAX_LONG (huge!) and
+      * {@link Approximation#ROW_COUNT_FOR_COUNT_ESTIMATION} = 10000.
+      * To be safe, the maximum iteration count is capped at 10, and an exception is thrown
+      * when this count is exceeded.
+      */
+    private LogicalPlan processCount(long rowCount) {
+        subPlanIterationCount += 1;
+        if (subPlanIterationCount > 10) {
+            throw new IllegalStateException("Approximation count iteration limit exceeded");
+        }
+        double sampleProbability = nextSubPlanSampleProbability;
         logger.debug("estimated number of rows reaching STATS (p=[{}]): [{}] rows", sampleProbability, rowCount);
         double newSampleProbability = Math.min(
             1.0,
-            sampleProbability * sampleRowCount(configuration.approximationSettings()) / Math.max(1, rowCount)
+            sampleProbability * sampleRowCount() / Math.max(1, rowCount)
         );
         if (newSampleProbability > SAMPLE_PROBABILITY_THRESHOLD) {
             // If the new sample probability is large, run the original query.
             logger.debug("using original plan (too few rows)");
-            return substituteSampleProbability(logicalPlan, 1.0);
+            nextSubPlanSampleProbability = null;
+            return ApproximationPlan.substituteSampleProbability(logicalPlan, 1.0);
         } else if (rowCount <= ROW_COUNT_FOR_COUNT_ESTIMATION / 2) {
             // Not enough rows are sampled yet; increase the sample probability and try again.
+            nextSubPlanSampleProbability = Math.min(1.0, sampleProbability * ROW_COUNT_FOR_COUNT_ESTIMATION / Math.max(1, rowCount));;
             return logicalPlan;
         } else {
-            return substituteSampleProbability(logicalPlan, newSampleProbability);
+            // A good sample probability is found; run the approximation plan.
+            nextSubPlanSampleProbability = null;
+            return ApproximationPlan.substituteSampleProbability(logicalPlan, newSampleProbability);
+        }
+    }
+
+    /**
+     * Returns the target number of rows to sample for approximation.
+     */
+    private int sampleRowCount() {
+        if (settings.rows() != null) {
+            return settings.rows();
+        } else if (queryProperties.hasGrouping) {
+            return DEFAULT_ROW_COUNT_WITH_GROUPING;
+        } else {
+            return DEFAULT_ROW_COUNT_WITHOUT_GROUPING;
         }
     }
 
     /**
      * Returns the row count in the result and closes the result.
      */
-    private static long rowCount(Result countResult) {
+    private long rowCount(Result countResult) {
         assert countResult.pages().size() == 1;
         assert countResult.pages().getFirst().getBlockCount() == 1;
         assert countResult.pages().getFirst().getPositionCount() == 1;
@@ -731,508 +549,5 @@ public class Approximation {
         long rowCount = ((LongBlock) (countResult.pages().getFirst().getBlock(0))).getLong(0);
         countResult.pages().getFirst().close();
         return rowCount;
-    }
-
-    /**
-     * Returns a plan that approximates the original plan and computes confidence intervals.
-     * This approximation query consists of the following:
-     * <ul>
-     *     <li> Source command
-     *     <li> {@code SAMPLE} with the provided sample probability
-     *     <li> All commands before the {@code STATS} command
-     *     <li> {@code EVAL} adding a new column with random bucket IDs for each trial
-     *     <li> {@code STATS} command with:
-     *          <ul>
-     *              <li> {@code COUNT} to track the sample size
-     *              <li> Each aggregate function replaced by a sample-corrected version (if needed)
-     *              <li> {@link Approximation#TRIAL_COUNT} * {@link Approximation#BUCKET_COUNT} additional columns
-     *                   with a sampled values for each aggregate function, sample-corrected (if needed)
-     *          </ul>
-     *     <li> {@code FILTER} to remove all rows with a too small sample size
-     *     <li> All commands after the {@code STATS} command, modified to also process
-     *          the additional bucket columns where possible
-     *     <li> {@code EVAL} to compute confidence intervals for all fields with buckets
-     *     <li> {@code PROJECT} to drop all non-output columns
-     * </ul>
-     *
-     * As an example, the simple query:
-     * <pre>
-     *     {@code
-     *         FROM index
-     *             | EVAL x = 2*x
-     *             | STATS s = SUM(x) BY group
-     *             | EVAL t = s*s
-     *     }
-     * </pre>
-     * is rewritten to (prob=sampleProbability, T=trialCount, B=bucketCount):
-     * <pre>
-     *     {@code
-     *         FROM index
-     *             | SAMPLE prob
-     *             | EVAL x = 2*x
-     *             | EVAL bucketId = MV_APPEND(RANDOM(B), ... , RANDOM(B))  // T times
-     *             | STATS sampleSize = COUNT(*),
-     *                     s = SUM(x) / prob,
-     *                     `s$0` = SUM(x) / (prob/B)) WHERE MV_SLICE(bucketId, 0, 0) == 0
-     *                     ...,
-     *                     `s$T*B-1` = SUM(x) / (prob/B) WHERE MV_SLICE(bucketId, T-1, T-1) == B-1
-     *               BY group
-     *             | WHERE sampleSize >= 10
-     *             | EVAL t = s*s, `t$0` = `s$0`*`s$0`, ..., `t$T*B-1` = `s$T*B-1`*`s$T*B-1`
-     *             | EVAL `CONFIDENCE_INTERVAL(s)` = CONFIDENCE_INTERVAL(s, MV_APPEND(`s$0`, ... `s$T*B-1`), T, B, 0.90),
-     *                    `CONFIDENCE_INTERVAL(t)` = CONFIDENCE_INTERVAL(t, MV_APPEND(`t$0`, ... `t$T*B-1`), T, B, 0.90)
-     *             | KEEP s, t, `CONFIDENCE_INTERVAL(s)`, `CONFIDENCE_INTERVAL(t)`
-     *     }
-     * </pre>
-     */
-    public LogicalPlan approximationPlan() {
-        logger.debug("generating approximation plan");
-
-        // Whether of not the first STATS command has been encountered yet.
-        Holder<Boolean> encounteredStats = new Holder<>(false);
-
-        // The keys are the IDs of the fields that have buckets. Confidence intervals are computed
-        // for these fields at the end of the computation. They map to the list of buckets for
-        // that field.
-        Map<NameId, List<Alias>> fieldBuckets = new HashMap<>();
-
-        // For each sample-corrected expression, also keep track of the uncorrected expression.
-        // These are used when a division between two sample-corrected expressions is encountered.
-        // This results in the same value (because (expr1/prob) / (expr2/prob) == expr1/expr2),
-        // except that no round-off errors occur if either the numerator or denominator is an
-        // integer and rounded to that after sample-correction. The most common case is AVG, which
-        // is rewritten to AVG::double = SUM::double / COUNT::long.
-        Map<NameId, NamedExpression> uncorrectedExpressions = new HashMap<>();
-
-        LogicalPlan approximationPlan = logicalPlan.transformUp(plan -> {
-            if (encounteredStats.get() == false) {
-                if (plan instanceof Aggregate == false) {
-                    // Commands before the first STATS function should be left unchanged.
-                    return plan;
-                } else {
-                    // The first STATS function should be replaced by a sample-corrected STATS
-                    // and buckets (for computing confidence intervals).
-                    encounteredStats.set(true);
-                    return sampleCorrectedAggregateAndBuckets((Aggregate) plan, fieldBuckets, uncorrectedExpressions);
-                }
-            } else {
-                // After the STATS function, any processing of fields that have buckets, should
-                // also process the buckets, so that confidence intervals for the dependent fields
-                // can be computed.
-                return planIncludingBuckets(plan, fieldBuckets, uncorrectedExpressions);
-            }
-        });
-
-        // Add the confidence intervals for all fields with buckets.
-        approximationPlan = new Eval(Source.EMPTY, approximationPlan, getConfidenceIntervals(fieldBuckets));
-
-        // Drop all bucket fields and uncorrected fields from the output.
-        Set<Attribute> dropAttributes = Stream.concat(
-            fieldBuckets.values().stream().flatMap(List::stream),
-            uncorrectedExpressions.values().stream()
-        ).map(NamedExpression::toAttribute).collect(Collectors.toSet());
-
-        List<Attribute> keepAttributes = new ArrayList<>(approximationPlan.output());
-        keepAttributes.removeAll(dropAttributes);
-        approximationPlan = new Project(Source.EMPTY, approximationPlan, keepAttributes);
-        logger.debug("approximation plan:\n{}", approximationPlan);
-        return approximationPlan;
-    }
-
-    /**
-     * Replaces the aggregate by a sample-corrected aggregate and buckets, and
-     * filters out groups with a too small sample size. This means that:
-     * <pre>
-     *     {@code
-     *          STATS s = SUM(x) BY group
-     *     }
-     * </pre>
-     * is replaced by:
-     * <pre>
-     *     {@code
-     *          STATS sampleSize = COUNT(*),
-     *                s = SUM(x),
-     *                `s$0` = SUM(x) WHERE MV_SLICE(bucketId, 0, 0) == 0
-     *                ...,
-     *                `s$T*B-1` = SUM(x) / (prob/B) WHERE MV_SLICE(bucketId, T-1, T-1) == B-1
-     *          BY group
-     *          | WHERE sampleSize >= MIN_ROW_COUNT_FOR_RESULT_INCLUSION
-     *          | EVAL s = s / prob, `s$0` = `s$0` / (prob/B), `s$T*B-1` = `s$T*B-1` / (prob/B)
-     *          | DROP sampleSize
-     *      }
-     * </pre>
-     */
-    private LogicalPlan sampleCorrectedAggregateAndBuckets(
-        Aggregate aggregate,
-        Map<NameId, List<Alias>> fieldBuckets,
-        Map<NameId, NamedExpression> uncorrectedExpressions
-    ) {
-        Expression sampleProbability = new SampleProbability();
-        Expression bucketSampleProbability = new Div(Source.EMPTY, sampleProbability, Literal.integer(Source.EMPTY, BUCKET_COUNT));
-
-        Expression randomBucketId = new Random(Source.EMPTY, Literal.integer(Source.EMPTY, BUCKET_COUNT));
-        Expression bucketIds = randomBucketId;
-        for (int trialId = 1; trialId < TRIAL_COUNT; trialId++) {
-            bucketIds = new MvAppend(Source.EMPTY, bucketIds, randomBucketId);
-        }
-        // TODO: use theoretically non-conflicting names.
-        Alias bucketIdField = new Alias(Source.EMPTY, "$bucket_id", bucketIds);
-
-        // The aggregate functions in the approximation plan.
-        List<NamedExpression> bucketAggregates = new ArrayList<>();
-
-        // List of expressions that must be evaluated after the sampled aggregation.
-        // These consist of:
-        // - sample corrections (to correct counts/sums for sampling)
-        // - replace zero counts by NULLs (for confidence interval computation)
-        // - exact total row count if COUNT(*) is used (to avoid sampling errors there)
-        List<Alias> evals = new ArrayList<>();
-        List<NamedExpression> originalAggregates = new ArrayList<>();
-        List<Attribute> projections = new ArrayList<>();
-
-        for (NamedExpression aggOrKey : aggregate.aggregates()) {
-            if ((aggOrKey instanceof Alias alias && alias.child() instanceof AggregateFunction) == false) {
-                // This is a grouping key, not an aggregate function.
-                originalAggregates.add(aggOrKey);
-                projections.add(aggOrKey.toAttribute());
-                continue;
-            }
-
-            Alias aggAlias = (Alias) aggOrKey;
-            AggregateFunction aggFn = (AggregateFunction) aggAlias.child();
-
-            if (SUPPORTED_SINGLE_VALUED_AGGS.contains(aggFn.getClass())) {
-                // For the supported single-valued aggregations, add buckets with sampled
-                // values, that will be used to compute a confidence interval.
-                // For multivalued aggregations, confidence intervals do not make sense.
-                List<Alias> buckets = new ArrayList<>();
-                for (int trialId = 0; trialId < TRIAL_COUNT; trialId++) {
-                    for (int bucketId = 0; bucketId < BUCKET_COUNT; bucketId++) {
-                        Expression bucketIdFilter = new Equals(
-                            Source.EMPTY,
-                            new MvSlice(
-                                Source.EMPTY,
-                                bucketIdField.toAttribute(),
-                                Literal.integer(Source.EMPTY, trialId),
-                                Literal.integer(Source.EMPTY, trialId)
-                            ),
-                            Literal.integer(Source.EMPTY, bucketId)
-                        );
-                        Expression bucket = aggFn.withFilter(
-                            aggFn.hasFilter() == false ? bucketIdFilter : new And(Source.EMPTY, aggFn.filter(), bucketIdFilter)
-                        );
-                        Alias bucketAlias = new Alias(
-                            Source.EMPTY,
-                            aggOrKey.name() + "$bucket$" + (trialId * BUCKET_COUNT + bucketId),
-                            bucket
-                        );
-                        if (SAMPLE_CORRECTED_AGGS.contains(aggFn.getClass()) == false) {
-                            buckets.add(bucketAlias);
-                            bucketAggregates.add(bucketAlias);
-                            projections.add(bucketAlias.toAttribute());
-                        } else {
-                            Alias uncorrectedBucketAlias = new Alias(Source.EMPTY, bucketAlias.name() + "$uncorrected", bucket);
-                            uncorrectedExpressions.put(bucketAlias.id(), uncorrectedBucketAlias);
-                            bucketAggregates.add(uncorrectedBucketAlias);
-                            projections.add(uncorrectedBucketAlias.toAttribute());
-
-                            Expression uncorrectedBucket = uncorrectedBucketAlias.toAttribute();
-                            if (aggFn.equals(COUNT_ALL_ROWS)) {
-                                // For COUNT, no data should result in NULL, like in other aggregations.
-                                // Otherwise, the confidence interval computation breaks.
-                                uncorrectedBucket = new Case(
-                                    Source.EMPTY,
-                                    new Equals(Source.EMPTY, uncorrectedBucket, Literal.fromLong(Source.EMPTY, 0L)),
-                                    List.of(Literal.NULL, uncorrectedBucket)
-                                );
-                            }
-
-                            Expression correctedBucket = correctForSampling(uncorrectedBucket, bucketSampleProbability, null);
-                            Alias correctedBucketAlias = bucketAlias.replaceChild(correctedBucket);
-                            evals.add(correctedBucketAlias);
-                            projections.add(correctedBucketAlias.toAttribute());
-                            buckets.add(correctedBucketAlias);
-                        }
-                    }
-                }
-                fieldBuckets.put(aggOrKey.id(), buckets);
-            }
-
-            // Replace the original aggregation by a sample-corrected one if needed.
-            if (SAMPLE_CORRECTED_AGGS.contains(aggFn.getClass()) == false) {
-                originalAggregates.add(aggAlias);
-                projections.add(aggAlias.toAttribute());
-            } else {
-                Alias uncorrectedAggAlias = new Alias(aggAlias.source(), aggAlias.name() + "$uncorrected", aggFn);
-                uncorrectedExpressions.put(aggAlias.id(), uncorrectedAggAlias);
-                originalAggregates.add(uncorrectedAggAlias);
-                projections.add(uncorrectedAggAlias.toAttribute());
-
-                Expression correctedAgg = correctForSampling(
-                    uncorrectedAggAlias.toAttribute(),
-                    sampleProbability,
-                    fieldBuckets.get(aggOrKey.id())
-                );
-                evals.add(aggAlias.replaceChild(correctedAgg));
-                projections.add(aggAlias.toAttribute());
-            }
-        }
-
-        List<NamedExpression> aggregates = Stream.concat(originalAggregates.stream(), bucketAggregates.stream()).collect(Collectors.toList());
-
-        Alias sampleSize = null;
-        if (aggregate.groupings().isEmpty() == false) {
-            // Add the sample size per grouping to filter out groups with too few sampled rows.
-            sampleSize = new Alias(Source.EMPTY, "$sample_size", COUNT_ALL_ROWS);
-            aggregates.add(sampleSize);
-            originalAggregates.add(sampleSize);
-        }
-
-        // Add the bucket ID, do the aggregations (sampled corrected, including the buckets),
-        // and filter out rows with too few sampled values.
-        LogicalPlan plan = new Eval(Source.EMPTY, aggregate.child(), List.of(bucketIdField));
-
-        plan = new SampledAggregate(aggregate.source(), plan, aggregate.groupings(), aggregates, originalAggregates, sampleProbability);
-
-        if (sampleSize != null) {
-            plan = new Filter(
-                Source.EMPTY,
-                plan,
-                new GreaterThanOrEqual(Source.EMPTY, sampleSize.toAttribute(), new SampleSizeThreshold())
-            );
-        }
-
-        plan = new Eval(Source.EMPTY, plan, evals);
-        return new Project(Source.EMPTY, plan, projections);
-    }
-
-    /**
-     * Corrects an aggregation function for random sampling.
-     * Some functions (like COUNT and SUM) need to be scaled up by the inverse of
-     * the sampling probability, while others (like AVG and MEDIAN) do not.
-     */
-    private static Expression correctForSampling(Expression expr, Expression sampleProbability, List<Alias> buckets) {
-        Expression correctedAgg = new Div(expr.source(), expr, sampleProbability);
-        correctedAgg = switch (expr.dataType()) {
-            case DOUBLE -> correctedAgg;
-            case LONG -> new ToLong(expr.source(), correctedAgg);
-            default -> throw new IllegalStateException("unexpected data type [" + expr.dataType() + "]");
-        };
-        if (buckets != null) {
-            // All buckets being null indicates that the query was executed
-            // exactly, hence no sampling correction must be applied.
-            List<Expression> rest = buckets.subList(1, buckets.size()).stream().map(Alias::toAttribute).collect(Collectors.toList());
-            correctedAgg = new Case(
-                Source.EMPTY,
-                new IsNull(Source.EMPTY, new Coalesce(Source.EMPTY, buckets.getFirst().toAttribute(), rest)),
-                List.of(expr, correctedAgg)
-            );
-        }
-        return correctedAgg;
-    }
-
-    /**
-     * Returns a plan that also processes the buckets for fields that have them.
-     * Luckily, there's only a limited set of commands that have to do something
-     * with the buckets: EVAL, PROJECT and MV_EXPAND.
-     */
-    private LogicalPlan planIncludingBuckets(
-        LogicalPlan plan,
-        Map<NameId, List<Alias>> fieldBuckets,
-        Map<NameId, NamedExpression> uncorrectedExpressions
-    ) {
-        return switch (plan) {
-            case Eval eval -> evalIncludingBuckets(eval, fieldBuckets, uncorrectedExpressions);
-            case Project project -> projectIncludingBuckets(project, fieldBuckets);
-            case MvExpand mvExpand -> mvExpandIncludingBuckets(mvExpand, fieldBuckets);
-            default -> plan;
-        };
-    }
-
-    /**
-     * For EVAL, if any of the evaluated expressions depends on a field with buckets,
-     * create corresponding expressions for the buckets as well. If the results are
-     * non-numeric or multivalued, though, confidence intervals don't apply anymore,
-     * and the buckets not computed.
-     */
-    private LogicalPlan evalIncludingBuckets(
-        Eval eval,
-        Map<NameId, List<Alias>> fieldBuckets,
-        Map<NameId, NamedExpression> uncorrectedExpressions
-    ) {
-        List<Alias> fields = new ArrayList<>(eval.fields());
-        for (Alias field : eval.fields()) {
-            // Don't create buckets for non-numeric or multivalued fields.
-            if (field.dataType().isNumeric() == false
-                || field.child().anyMatch(expr -> MULTIVALUED_OUTPUT_FUNCTIONS.contains(expr.getClass()))) {
-                continue;
-            }
-            // If any of the field's dependencies has buckets, create buckets for this field as well.
-            if (field.child().anyMatch(e -> e instanceof NamedExpression ne && fieldBuckets.containsKey(ne.id()))) {
-                List<Alias> buckets = new ArrayList<>();
-                for (int bucketId = 0; bucketId < TRIAL_COUNT * BUCKET_COUNT; bucketId++) {
-                    final int finalBucketId = bucketId;
-                    Expression bucket = field.child()
-                        .transformDown(
-                            e -> e instanceof NamedExpression ne && fieldBuckets.containsKey(ne.id())
-                                ? fieldBuckets.get(ne.id()).get(finalBucketId).toAttribute()
-                                : e
-                        );
-                    buckets.add(new Alias(Source.EMPTY, field.name() + "$" + bucketId, bucket));
-                }
-                fields.addAll(buckets);
-                fieldBuckets.put(field.id(), buckets);
-            }
-        }
-        // For each division of two sample-corrected expressions, replace it by
-        // a division of the corresponding uncorrected expressions.
-        for (int i = 0; i < fields.size(); i++) {
-            Alias field = fields.get(i);
-            fields.set(i, field.replaceChild(field.child().transformUp(e -> {
-                if (e instanceof Div div
-                    && div.left() instanceof NamedExpression left
-                    && uncorrectedExpressions.containsKey(left.id())
-                    && div.right() instanceof NamedExpression right
-                    && uncorrectedExpressions.containsKey(right.id())) {
-                    return new Div(
-                        e.source(),
-                        uncorrectedExpressions.get(left.id()).toAttribute(),
-                        uncorrectedExpressions.get(right.id()).toAttribute(),
-                        div.dataType()
-                    );
-                }
-                return e;
-            })));
-        }
-        return new Eval(Source.EMPTY, eval.child(), fields);
-    }
-
-    /**
-     * For PROJECT, if it renames a field with buckets, add the renamed field
-     * to the map of fields with buckets.
-     */
-    private LogicalPlan projectIncludingBuckets(Project project, Map<NameId, List<Alias>> fieldBuckets) {
-        for (NamedExpression projection : project.projections()) {
-            if (projection instanceof Alias alias
-                && alias.child() instanceof NamedExpression named
-                && fieldBuckets.containsKey(named.id())) {
-                fieldBuckets.put(alias.id(), fieldBuckets.get(named.id()));
-            }
-        }
-
-        // When PROJECT keeps a field with buckets, also keep the buckets.
-        List<NamedExpression> projections = null;
-        for (NamedExpression projection : project.projections()) {
-            if (fieldBuckets.containsKey(projection.id())) {
-                if (projections == null) {
-                    projections = new ArrayList<>(project.projections());
-                }
-                for (Alias bucket : fieldBuckets.get(projection.id())) {
-                    projections.add(bucket.toAttribute());
-                }
-            }
-        }
-        if (projections != null) {
-            project = project.withProjections(projections);
-        }
-        return project;
-    }
-
-    /**
-     * Fields with buckets are always single-valued, so expanding them doesn't
-     * do anything and the buckets of the expanded field are the same as those
-     * of the target field.
-     */
-    private LogicalPlan mvExpandIncludingBuckets(MvExpand mvExpand, Map<NameId, List<Alias>> fieldBuckets) {
-        if (fieldBuckets.containsKey(mvExpand.target().id())) {
-            fieldBuckets.put(mvExpand.expanded().id(), fieldBuckets.get(mvExpand.target().id()));
-        }
-        return mvExpand;
-    }
-
-    /**
-     * Returns the confidence interval and certified fields for the fields.
-     * This is the expression:
-     * <pre>
-     *     {@code
-     *         CONFIDENCE_INTERVAL(s, MV_APPEND(`s$0`, ... `s$T*B-1`), T, B, 0.90)
-     *     }
-     * </pre>
-     * for each field {@code s} that has buckets. The output of {@code CONFIDENCE_INTERVAL}
-     * is separated into two fields: the confidence interval itself, and a certified field.
-     */
-    private List<Alias> getConfidenceIntervals(Map<NameId, List<Alias>> fieldBuckets) {
-        Expression constNaN = new Literal(Source.EMPTY, Double.NaN, DataType.DOUBLE);
-        Expression trialCount = Literal.integer(Source.EMPTY, TRIAL_COUNT);
-        Expression bucketCount = Literal.integer(Source.EMPTY, BUCKET_COUNT);
-        Expression confidenceLevel = Literal.fromDouble(Source.EMPTY, confidenceLevel());
-
-        // Compute the confidence interval for all output fields that have buckets.
-        List<Alias> confidenceIntervalsAndCertified = new ArrayList<>();
-        for (Attribute output : logicalPlan.output()) {
-            if (fieldBuckets.containsKey(output.id())) {
-                List<Alias> buckets = fieldBuckets.get(output.id());
-                // Collect a multivalued expression with all bucket values, and pass that to the
-                // confidence interval computation. Whenever the bucket value is null, replace it
-                // by NaN, because multivalued fields cannot have nulls.
-                // This is a bit of a back, because ES|QL generally does not support NaN values,
-                // but these values stay inside here and the confidence interval computation, and
-                // never reach the user.
-                // TODO: don't use NaNs, perhaps when nulls in multivalued are supported, see:
-                // https://github.com/elastic/elasticsearch/issues/141383
-                Expression bucketsMv = null;
-                for (int i = 0; i < TRIAL_COUNT * BUCKET_COUNT; i++) {
-                    Expression bucket = buckets.get(i).toAttribute();
-                    if (output.dataType() != DataType.DOUBLE) {
-                        bucket = new ToDouble(Source.EMPTY, bucket);
-                    }
-                    bucket = new Case(Source.EMPTY, new IsNotNull(Source.EMPTY, bucket), List.of(bucket, constNaN));
-                    if (bucketsMv == null) {
-                        bucketsMv = bucket;
-                    } else {
-                        bucketsMv = new MvAppend(Source.EMPTY, bucketsMv, bucket);
-                    }
-                }
-                Expression outputDouble = output.dataType() == DataType.DOUBLE ? output : new ToDouble(Source.EMPTY, output);
-                Expression confidenceInterval = new ConfidenceInterval(
-                    Source.EMPTY,
-                    outputDouble,
-                    bucketsMv,
-                    trialCount,
-                    bucketCount,
-                    confidenceLevel
-                );
-                confidenceInterval = switch (output.dataType()) {
-                    case DOUBLE -> confidenceInterval;
-                    case INTEGER -> new ToInteger(Source.EMPTY, confidenceInterval);
-                    case LONG -> new ToLong(Source.EMPTY, confidenceInterval);
-                    default -> throw new IllegalStateException("unexpected data type [" + output.dataType() + "]");
-                };
-                confidenceIntervalsAndCertified.add(
-                    new Alias(
-                        Source.EMPTY,
-                        "CONFIDENCE_INTERVAL(" + output.name() + ")",
-                        new MvSlice(Source.EMPTY, confidenceInterval, Literal.integer(Source.EMPTY, 0), Literal.integer(Source.EMPTY, 1))
-                    )
-                );
-                confidenceIntervalsAndCertified.add(
-                    new Alias(
-                        Source.EMPTY,
-                        "CERTIFIED(" + output.name() + ")",
-                        new GreaterThanOrEqual(
-                            Source.EMPTY,
-                            new MvSlice(
-                                Source.EMPTY,
-                                confidenceInterval,
-                                Literal.integer(Source.EMPTY, 2),
-                                Literal.integer(Source.EMPTY, 2)
-                            ),
-                            Literal.fromDouble(Source.EMPTY, 0.5)
-                        )
-                    )
-                );
-            }
-        }
-        return confidenceIntervalsAndCertified;
     }
 }
